extract logic from Unit component as useUnit hook
